Tighten a bit (psalm)

+++ src/Base83.php
@@ -32,7 +32,7 @@ class Base83 {
     public static function decode(string $hash): int {
         $result = 0;
         foreach (str_split($hash) as $char) {
-            $result = $result * self::BASE + array_search($char, self::ALPHABET, true);
+            $result = $result * self::BASE + (int) array_search($char, self::ALPHABET, true);
         }
         return intval ($result);
     }
